Rename ModelUtils to AlertingUtils for more general use
