+++ alerting/src/main/kotlin/com/amazon/opendistroforelasticsearch/alerting/model/destination/email/Email.kt
@@ -15,7 +15,7 @@
 
 package com.amazon.opendistroforelasticsearch.alerting.model.destination.email
 
-import com.amazon.opendistroforelasticsearch.alerting.util.ModelUtils
+import com.amazon.opendistroforelasticsearch.alerting.util.isValidEmail
 import org.elasticsearch.common.io.stream.StreamInput
 import org.elasticsearch.common.io.stream.StreamOutput
 import org.elasticsearch.common.io.stream.Writeable
@@ -117,7 +117,7 @@ data class Recipient(
             RecipientType.EMAIL_GROUP -> requireNotNull(emailGroupID) { "Email group ID is null" }
             RecipientType.EMAIL -> {
                 requireNotNull(email) { "Email is null" }
-                require(ModelUtils.isValidEmail(email)) { "Invalid email" }
+                require(isValidEmail(email)) { "Invalid email" }
             }
         }
     }
